plugins.v2/clouddrivedisk/requirements.txt
grpcio~=1.78.1
protobuf~=5.29.6
httpx[http2]
